+++ src/components/DebouncedInput.tsx
@@ -1,0 +1,32 @@
+import React, { useState, useEffect } from 'react';
+
+// A debounced input react component
+function DebouncedInput({
+    value: initialValue,
+    onChange,
+    debounce = 500,
+    ...props
+  }: {
+    value: string | number
+    onChange: (value: string | number) => void
+    debounce?: number
+  } & Omit<React.InputHTMLAttributes<HTMLInputElement>, 'onChange'>) {
+    const [value, setValue] = useState(initialValue)
+  
+    useEffect(() => {
+      setValue(initialValue)
+    }, [initialValue])
+  
+    useEffect(() => {
+      const timeout = setTimeout(() => {
+        onChange(value)
+      }, debounce)
+  
+      return () => clearTimeout(timeout)
+    }, [value])
+  
+    return (
+      <input {...props} value={value} onChange={e => setValue(e.target.value)} />
+    )
+  }
+  export default DebouncedInput;
